dev-0524 Moved StartTestCaseInfo and EndTestCaseInfo to the message-collector due to type import issues.
