+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Please see [CONTRIBUTING.md](https://github.com/LynxWall/cucumber-js-tsflow/blob/master/CONTRIBUTE.md) on how to contribute to cucumber-tsflow.
 
+## [6.5.4]
+
+### Fixed
+
+- Moved StartTestCaseInfo and EndTestCaseInfo to the message-collector due to type import issues.
+
 ## [6.5.3]
 
 ### Fixed
